Use style sizing to fix issues with Edge + flex + canvas.
Additionally render the canvas at a multiplier of the device pixel
ratio, so we get hidpi images for supported screens. The fixed CSS
sizing allows us to do that.

The problem was that on Edge, flexbox apparently can grow canvas
elements, whereas on Chrome/FF the canvas width/height is respected. No
problem, we can set the width/height in CSS to fix it, which
conveniently allows us to also fix the low-res-on-retina-screens problem
at the same time.