log workspace update errors
Workspace update errors can be partially recovered from. This allows
the gui to "successfully" load on projects that pass vm validation and
loading but throw an error in scratch-blocks. Logging the error instead
of throwing it allows the gui and vm to finish updating for the
workspace change. This allows the VM to behave normally even if the
scratch-blocks workspace rendering is incomplete.