Handle ?tutorial= query in gui
Moved the tutorial id handling into gui so it works when gui is embedded in www as well as standalone.

This seemed better than adding the same query handling to www, in order to pass an id to gui.

Note, the `?tutorial=all` handling remains in app-state-hoc. Currently it’s only needed in standalone mode for Hour of Code. It’s easy to move into the gui container if we do want it in gui by default.