Make font inlining a string operation instead of DOM
This makes loading vector costumes quite a bit faster, since it removes the need to clone the SVG dom and insert large nodes into it, which is a very slow operation. Replace with a simple string manipulation equivalent. On my machine this makes the `toString` operation take about 15-25ms less time.