Allow drop event in sortableHOC even when there are no items to sort.
This is used because the sortable drop is also used for general dropping as well as sorting drop, so you need to be able to drop into a sortable list that does not contain any items.

Make the corresponding reorder code be more careful to check if there is a newIndex.