feat: make the account menu more configurable
Previously the menu determined its own state based on the redux session.
Since we don't use that in NGP we want to be able to configure the menu
in a "dumber" way - provide the necessary parameters so that the editor
only gets the values it needs to show/use and doesn't try to determine it
itself.

Part of the story is that the old session data model is different from
the new one. Another part is that the `username` is no longer a unique
identifier (as in a classroom context it may be the actual person name).
And a third part of the story is that some functionality is not (yet)
applicable to NGP, for example "My stuff", logging in from the editor,
account settings, etc.

So we introduce the AccountMenuOptions interface which can be passed
as a prop from the GUI container. If that prop is present - the editor
will use it to determine what to show in the account menu.
If it's not present - then nothing changes - logic is as it was before,
so that we're backwards-compatible with scratch-www.